+++ hammerblade/torch/kernel/kernel_common.hpp
@@ -27,6 +27,7 @@
 #include "hb_tensor.hpp"
 #include <hb_assert.hpp>
 #include <hb_tiled_for.hpp>
+#include <hb_blocked_for.hpp>
 #include <hb_spatial_for.hpp>
 #include <hb_common.hpp>
